CollegeSurvey Existing Exam Check Shell
Correct Functions created for checking if existing SAT/ACT scores exist and rendering correct display in the event they are. Correct json backend name for act and sat scores needs to be updated. Also needs to be tested with a functional backend, tested for function without use of fetch function.